Phase 2: Stat Modifiers (Speed & Agility)
Implemented stat-based speed modifiers using linear scaling from neutral point (50).
Replaced old modifier logic with clean, configuration-driven calculator.

Changes:
- Implemented CalculateStatModifiers in SpeedModifierCalculator
  - Speed modifier: ±10% range (0.002 per point from 50)
  - Agility modifier: ±5% range (0.001 per point from 50)
  - Linear formulas: 1.0 + ((stat - 50) * modifierPerPoint)
- Refactored RaceService.UpdateHorsePosition to use new calculator
  - Reduced method from ~60 lines to ~30 lines
  - Clear modifier pipeline: Stats → Env → Phase → Random
- Added comprehensive stat modifier tests (9 new tests)
- Added RaceService integration test for stat modifiers

Test Results: 73/73 tests passing
Follows TDD Red-Green-Refactor cycle with vertical slice delivery

+++ TripleDerby.Core/Racing/ModifierContext.cs
@@ -14,8 +14,8 @@ namespace TripleDerby.Core.Racing;
 /// <param name="RaceSurface">Track surface type (dirt, turf, artificial)</param>
 /// <param name="RaceFurlongs">Race distance in furlongs</param>
 public record ModifierContext(
-    int CurrentTick,
-    int TotalTicks,
+    short CurrentTick,
+    short TotalTicks,
     Horse Horse,
     ConditionId RaceCondition,
     SurfaceId RaceSurface,
